Improve disk drive noises.
The requested step time is now taken into account, so it sounds much
tidier. Still not always quite right, as the requested step time can
be different from the available samples - but maybe this could be
fixable?

See #2.

+++ src/beeb/include/beeb/1770.h
@@ -40,10 +40,10 @@ class WD1770Handler {
     virtual bool IsTrack0() = 0;
 
     // Step out (towards track 0).
-    virtual void StepOut(int step_rate_ms) = 0;
+    virtual void StepOut(uint8_t step_rate_ms) = 0;
 
     // Step in (towards track N).
-    virtual void StepIn(int step_rate_ms) = 0;
+    virtual void StepIn(uint8_t step_rate_ms) = 0;
 
     // Spin the drive up. The motor is assumed to switch on.
     virtual void SpinUp() = 0;
